Fixed topbar + js
Fixed the topbar and nav section to be more responsive.

Tried to apply a carousel of latest posts to the index page. JS is currently printing "An error has occurred" to the page, which means the error catching is working.

Need to figure out what the error is and fix it.