Fixed problem with js generated html
Found out and fixed the problems with latest posts list being generated through JS, fixed it.

Added similarly generated HTML to about us-page to add author information.

Discovered only three of five authors have been added to the API - need to fix that sooner.

Did some stylizing to the generated html. Need to stylize it further and make sure it is responsive and works as a slider/carousel instead of a list.