Added content to post.html through post.js
Programmed post.js to generate html for post.html to display posts as you click on the links provided from the post list or latest post section of the index page.

Added CSS to have the post look better.

Added a link to the logo and made it appear black instead of the default link colour (red).