Added JS, CSS and generatet HTML to several pages
Fixed some errors I was having with my js files to provide more interactivity and functionality to several pages.

Had more js generated HTML to some pages to make the pages more like I was envisioning.

Did changes to CSS in accordance with the changes in js and html.

Added an image used for a new post that's being displayed on the site.