Recieved feedback and made changes based on that.
I recieved some feedback that the text should be framed somehow, and that the background should have a texture.

I took a 100px*100px artboard in Photoshop, added white background and took the red shade from the site design altered it and added little dots here and there for texture. I also added backgrounds to all containers with text.

I changed the border settings for most elements to see what I thought. I am going to leave it at that for a while and see if I still agree in one-two weeks.

Added small changes here and there.