Added images for posts + credit text file.
Added several images for new posts.

Added a text file to make note of who to credit for all visuals used for this project.

Made changes to the user images at the "about" page to load the biggest versions to the page and display them slightly bigger.

Added line breaks to index.js and post-list.js to have the generated html seem tidier.

I removed the "image modal" comment from post.js, as the modal codes was added to the try catch method.

Added a button for loading more posts in post-list.html.

Changed the api link used in post-list.js to have it be able to generate more than just the ten latest posts.