Tried to fix problems in js
Tried to fix the following:
- Allowing to validate the form, while also being able to send it if all fields were valid. Did not work.
- Loading 10 more posts in post list on clicking the "Load more" button. Did not work.
- Enabling sharing when clicking the social media buttons on each post. Did not work.

I also fixed some issues with the mobile media screen; the nav expanded beyond the header when activated, and the descriptive content containers on each page expanded too far to the right. I changed the width and padding of the nav, as well as the padding for the content containers.

Adding validation also made some strange changes to the form where it would not be centered anymore. I fixed it by removing the display: block feature on the validation elements and doing some changes to the flexbox features for the elements.