Added image for blog post, removed disable load button
Tried some solutions to make the "Load More" button in the blog post list not load more posts when there's no more posts to load in the database, since I could not make it work.

Also added a post to the database, and added an image to the image folder to add to the post.